Fix type of global var states
In the example index.html a global var named "states" is defined to collect states from the underlying iobrocker host. Unfortunately it is defined with an empty array which will later be overwritten with an resulting object as returned by the callback of "servConn.getStates". As this is an example the type should be correct.